+++ CHANGELOG.md
@@ -27,6 +27,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Proxy no longer fails on a response compressed with an unsupported algorithm - [#72](https://github.com/hivesolutions/netius/issues/72)
 * Proxy no longer truncates a compressed response that it decompresses before sending - [#72](https://github.com/hivesolutions/netius/issues/72)
 * Compression is no longer shared between the parallel requests of an HTTP/2 client - [#72](https://github.com/hivesolutions/netius/issues/72)
+* Memory used to compress an HTTP/2 request is now released as soon as the request ends - [#72](https://github.com/hivesolutions/netius/issues/72)
+* Proxy no longer discards a response when the back-end drops the connection halfway - [#72](https://github.com/hivesolutions/netius/issues/72)
 * DHT client no longer crashes when receiving a response with no matching pending request
 * DHT queries now send a valid node identifier so that peers can actually be discovered
 * DHT client now tolerates malformed responses and skips unroutable nodes instead of raising